% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapAUC}
\alias{bootstrapAUC}
\title{Bootstrap the AUC and compute a p-value for each profiled signature}
\usage{
bootstrapAUC(SE_scored, annotationData, signatureColNames,
  num.boot = 100, pb.show = TRUE)
}
\arguments{
\item{SE_scored}{A SummarizedExperiment object with genes in the rows,
and signature scores in the columns. There should also be a column indicating TB status.}

\item{annotationData}{A factor of annotation data, indicating TB status.}

\item{num.boot}{The number of times to bootstrap the data.}

\item{pb.show}{Logical for whether to show a progress bar while running code.}

\item{SignatureColNames}{The column names in the SE_scored colData that contain 
the signature score data.}
}
\value{
A list of length 3 returning a vector of p-values for a 2-sample t-test, bootstrapped AUC values,
and AUC value for using all scored values for all signatures specified in SignatureColNames.
}
\description{
Run bootstrapping of the AUC and derive the p-value for a t-test.
for all signatures tested on a given dataset.
}
