% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{compareBoxplots}
\alias{compareBoxplots}
\title{Create an array of boxplots for bootstrapped AUC values.}
\usage{
compareBoxplots(SE_scored, annotationData, signatureColNames,
  num.boot = 100, plot.col = 2, cex.axis = 0.7, cex = 0.25,
  name = "Boxplot Comparison of Signatures")
}
\arguments{
\item{SE_scored}{A SummarizedExperiment object with genes in the rows,
and signature scores in the columns. There should also be a column indicating TB status.}

\item{annotationData}{A vector of annotation data, typically indicating TB status.}

\item{num.boot}{The number of times to bootstrap the data.}

\item{plot.col}{}

\item{SignatureColNames}{The column names in the SE_scored colData that contain 
the signature score data.}
}
\value{

}
\description{
Present the results of AUC bootstrapping for the various scored signatures via boxplots.
}
