% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{deseq2_norm_rle}
\alias{deseq2_norm_rle}
\title{gene expression normalization, RLE normalization: relative log expression}
\usage{
deseq2_norm_rle(inputData)
}
\arguments{
\item{inputData}{a data.frame or matrix of gene expression count data. 
Required.}
}
\value{
Normalized count data
}
\description{
gene expression normalization, RLE normalization: relative log expression
}
\examples{
inputTest <- matrix(rnorm(1000), 100, 10, 
                    dimnames = list(paste0("gene", 1:100),
                                    paste0("sample", 1:10)))
res <- deseq2_norm_rle(inputTest)
}
