% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalModel.R
\name{Jacobsen_Sambarey_OriginalModel}
\alias{Jacobsen_Sambarey_OriginalModel}
\title{Train original model for gene signatures Jacobsen_3 and Sambarey_HIV_10.}
\usage{
Jacobsen_Sambarey_OriginalModel(dat_list, dat_test_sig)
}
\arguments{
\item{dat_list}{A list contains training data and disease status outcomes from
the discovery data of corresponding gene signatures.}

\item{dat_test_sig}{A data frame contains corresponding biomarker gene features.
Used to evaluate the performance of biomarker using its original models.}
}
\value{
The predicted score for each sample in the test study.
}
\description{
Train original model for gene signatures Jacobsen_3 and Sambarey_HIV_10.
}
