% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalModel.R
\name{LeongOriginalModel}
\alias{LeongOriginalModel}
\title{Train original model for gene signatures Leong_24 and Leong_RISK_29.}
\usage{
LeongOriginalModel(dat_list, dat_test_sig)
}
\arguments{
\item{dat_list}{A list contains training data and disease status outcomes from
the discovery data of corresponding gene signatures.}

\item{dat_test_sig}{A data frame contains correpsonding biomarker gene features.
Used to evaluate the performance of biomarker using its original models.}
}
\value{
The predicted score for each sample in the test study.
}
\description{
Train original model for gene signatures Leong_24 and Leong_RISK_29.
}
