% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{LOOAUC_simple_multiple_noplot_one_df}
\alias{LOOAUC_simple_multiple_noplot_one_df}
\title{LOOCV with logistic regression}
\usage{
LOOAUC_simple_multiple_noplot_one_df(df, targetVec)
}
\arguments{
\item{df}{a data.frame of gene expression count data.
Required.}

\item{targetVec}{binary vector indicating y}
}
\value{
AUC from LOOCV
}
\description{
LOOCV with logistic regression
}
\examples{
inputTest <- matrix(rnorm(1000), 100, 10,
                    dimnames = list(paste0("gene", 1:100),
                                    paste0("sample", 1:10)))
inputTest <- as.data.frame(inputTest)
targetVec <- sample(c(0,1), replace=TRUE, size=10)
res <- LOOAUC_simple_multiple_noplot_one_df(inputTest, targetVec)
}
