% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OriginalTrainingData}
\alias{OriginalTrainingData}
\title{Discovery datasets for correpsonding gene signatures}
\format{
list
}
\source{
See \code{?TBsignatures} for reference information.
}
\usage{
OriginalTrainingData
}
\description{
Discovery datasets for correpsonding gene signatures
}
\examples{
data("OriginalTrainingData")
}
\keyword{datasets}
