% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{runTBsigProfiler}
\alias{runTBsigProfiler}
\title{Run TB Signature Profiling}
\usage{
runTBsigProfiler(input, useAssay = NULL, signatures = NULL,
  algorithm = c("GSVA", "ssGSEA", "ASSIGN"),
  combineSigAndAlgorithm = FALSE, assignDir = NULL,
  outputFormat = NULL, parallel.sz = 0, ASSIGNiter = 1e+05,
  ASSIGNburnin = 50000)
}
\arguments{
\item{input}{Input data as a SummarizedExperiment, data.frame, or matrix.
Required}

\item{useAssay}{When using SummarizedExperiment style input, the assay to
use for signature profiling. Required for SummarizedExperiment}

\item{signatures}{A list of signatures to run. If signatures is NULL, the
default set of signatures from the included TBsignatures data set is run.
For details run ?TBsignatures.}

\item{algorithm}{A list of algorithms to run. The default is GSVA, ssGSEA,
and ASSIGN.}

\item{combineSigAndAlgorithm}{Logical. If TRUE, the row names will be in
the form <signature>_<algorithm>. If FALSE, there will be a column named
'algorithm' that lists which algorithm is used. If NULL, and one algorithm
was used, the algorithm will not be listed. The default is false.}

\item{assignDir}{If using ASSIGN, a directory to save intermediate ASSIGN
results. The default is NULL and intermediate results will not be saved.}

\item{outputFormat}{Output data format. Can be SummarizedExperiment, matrix,
or data.frame. The default is the same as input.}

\item{parallel.sz}{Number of processors to use when doing the calculations in
parallel for GSVA and ssGSEA. If 0, all cores are used. The default is 0.}

\item{ASSIGNiter}{The number of iterations in the ASSIGN MCMC. The default is
100000.}

\item{ASSIGNburnin}{The number of burn-in iterations for ASSIGN. These
iterations are discarded when computing the posterior means of the model
parameters. The default is 50000.}
}
\value{
A data.frame of signature profiling results.
}
\description{
Run profiling using a set of signatures and profiling algorithms
}
\examples{
#create some toy data to test Predict29 signature, with 5 samples with low
#and five samples with high expression of the signatures genes.
mat_testdata <- rbind(matrix(c(rnorm(145), rnorm(145) + 5), 29, 10,
                             dimnames = list(TBsignatures$Predict29,
                             paste0("sample", 1:10))),
                      matrix(rnorm(1000), 100, 10,
                             dimnames = list(paste0("gene", 1:100),
                             paste0("sample", 1:10))))
res <- runTBsigProfiler(mat_testdata, algorithm = "GSVA", parallel.sz = 1)
res["Predict29", ]
}
