% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{signatureBoxplot}
\alias{signatureBoxplot}
\title{Plot a Boxplot of Signature Genes}
\usage{
signatureBoxplot(inputData, annotationData, signatureColNames,
  annotationColName, name = "Signatures", scale = FALSE)
}
\arguments{
\item{inputData}{Either a SummarizedExperiment object that contains the
signature data and annotation as colData columns, or a data.frame or matrix
of signature data. Required.}

\item{annotationData}{If inputData is a data.frame or matrix of signature
data, a data.frame or matrix of annotation data.}

\item{signatureColNames}{If inputData is a SummarizedExperiment, the column
names in colData that contain the signature data.}

\item{annotationColName}{If inputData is a SummarizedExperiment, the column
name in colData to use to make the boxplot.}

\item{name}{The name of the boxplot. The default is "Signatures".}

\item{scale}{Scale the signature data. The default is FALSE.}
}
\value{
A ggplot2 boxplot of the signature data using the annotation
provided.
}
\description{
Plot a Boxplot of Signature Genes
}
\examples{
mat_testdata <- rbind(matrix(c(rnorm(145), rnorm(145) + 5), 29, 10,
                             dimnames = list(TBsignatures$Predict29,
                             paste0("sample", 1:10))),
                      matrix(rnorm(1000), 100, 10,
                             dimnames = list(paste0("gene", 1:100),
                             paste0("sample", 1:10))))
res <- runTBsigProfiler(mat_testdata, algorithm = "GSVA", parallel.sz = 1)
}
