% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{signatureROCplot}
\alias{signatureROCplot}
\title{Create an array of ROC plots to compare signatures.}
\usage{
signatureROCplot(inputData, annotationData, signatureColNames,
  annotationColName, scale = FALSE)
}
\arguments{
\item{inputData}{Either a SummarizedExperiment object that contains the
signature data and annotation as colData columns, or a data.frame or matrix
of signature data. Required.}

\item{annotationData}{If inputData is a data.frame or matrix of signature
data, a data.frame or matrix of annotation data.}

\item{signatureColNames}{If inputData is a SummarizedExperiment, the column
names in colData that contain the signature data.}

\item{annotationColName}{If inputData is a SummarizedExperiment, the column
name in colData that contains the annotation data.}

\item{scale}{Scale the signature data. The default is FALSE.}
}
\value{
An array of ROC plots.
}
\description{
Create an array of ROC plots to compare signatures.
}
