% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalModel.R
\name{subsetGeneSet}
\alias{subsetGeneSet}
\title{Filter gene expression value matrix based on certain gene sets.}
\usage{
subsetGeneSet(
  theObject,
  gene_set,
  useAssay = 1,
  obtainDiagnosis = FALSE,
  annotationColName = NULL
)
}
\arguments{
\item{theObject}{A SummarizedExperiment object that has been prestored in the
RData file: OriginalTrainingData.}

\item{gene_set}{A character vector that includes gene symbols for gene signatures.}

\item{useAssay}{A character string or an integer specifying the assay in the \code{theObject}
that will be selected. Default is 1, indicating the first assay of the SummarizedExperiment object.}

\item{obtainDiagnosis}{Boolean. Usually used to create training data if TRUE. Default is FALSE}

\item{annotationColName}{A character string specifying the column name of disease status.
Only used when creating training data. Default is NULL.}
}
\value{
A matrix with selected gene expression value if \code{obtainDiagnosis} is FALSE.
else, return a list contains the selected gene expression value and diagnosis results for each observation.
}
\description{
A function used to subset gene expression value matrix based on certain gene sets.
}
