% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalModel.R
\name{SulimanOriginalModel}
\alias{SulimanOriginalModel}
\title{Train original model gene signature Suliman_RISK_4.}
\usage{
SulimanOriginalModel(dat_list, dat_test_sig)
}
\arguments{
\item{dat_list}{A \code{list} contains training data and disease status outcomes from
the discovery data of corresponding gene signatures.}

\item{dat_test_sig}{A \code{data frame} contains corresponding gene sets from the \code{input}.}
}
\value{
The predicted score for each sample in the test study.
}
\description{
Train original model gene signature Suliman_RISK_4.
}
