% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalModel.R
\name{svm_OriginalModel}
\alias{svm_OriginalModel}
\title{Train original model for gene signatures Bloom_OD_144 and Zak_RISK_16.}
\usage{
svm_OriginalModel(dat_list, dat_test_sig)
}
\arguments{
\item{dat_list}{A \code{list} contains training data and disease status outcomes from
the discovery data of corresponding gene signatures.}

\item{dat_test_sig}{A \code{data frame} contains corresponding gene sets from the \code{input}.}
}
\value{
The predicted score for each sample in the test study.
}
\description{
Train original model for gene signatures Bloom_OD_144 and Zak_RISK_16.
}
