% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{tableAUC}
\alias{tableAUC}
\title{Create an interactive table of results for the bootstrapped AUC and p-values for a scored dataset.}
\usage{
tableAUC(SE_scored, annotationData, signatureColNames, num.boot,
  pb.show = TRUE)
}
\arguments{
\item{SE_scored}{A SummarizedExperiment object with genes in the rows,
and signature scores in the columns. There should also be a column indicating TB status.}

\item{annotationData}{A vector of annotation data, typically indicating TB status.}

\item{num.boot}{The number of times to bootstrap the data.}

\item{pb.show}{Logical for whether to show a progress bar while running code.}

\item{SignatureColNames}{The column names in the SE_scored colData that contain 
the signature score data.}
}
\value{

}
\description{
Collect the results of bootstrapping and t-tests for a scored gene expression dataset
and present using a JavaScript table with an R interface.
}
