% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TB_hiv}
\alias{TB_hiv}
\title{TB Example dataset with TB/HIV data}
\format{SummarizedExperiment}
\usage{
TB_hiv
}
\description{
An example dataset containing the gene expression and metadata in a SummarizedExperiment object for 31 subjects
with HIV and/or Tuberculosis diseases. Information on subject infection status can be accessed with \code{TB_hiv$Disease}.
Samples with both TB and HIV contamination are marked as \code{tb_hiv}, while samples
with HIV and no TB are marked as \code{hiv_only}.
}
\examples{
data("TB_hiv")
}
\keyword{datasets}
