% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TB_indian}
\alias{TB_indian}
\title{TB Example dataset with Indian population data}
\format{SummarizedExperiment}
\usage{
TB_indian
}
\description{
An example dataset containing the gene expression and metadata in a SummarizedExperiment object for an Indian 
population. Active TB contamination of the 44 subjects is denoted for each as a "1"(active) or "0" (latent/not present),
and can be accessed via \code{TB_indian$label}. The SummarizedExperiment object 
contains 2 assays (counts and log(counts)), and the column names give the unique
subject identification number along with the subject's gender.
}
\examples{
data("TB_indian")
}
\keyword{datasets}
