% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiler-package.R
\docType{package}
\name{TBSignatureProfiler-package}
\alias{TBSignatureProfiler}
\alias{TBSignatureProfiler-package}
\title{TBSignatureProfiler: Profile RNA-Seq Data Using TB Pathway Signatures}
\description{
Signatures of TB progression, TB disease, and other TB disease
    states have been created. This package makes it easy to profile RNA-Seq data
    using these signatures and common signature profiling tools including
    ASSIGN, GSVA, and ssGSEA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dfjenkins3/TBSignatureProfiler}
  \item Report bugs at \url{https://github.com/dfjenkins3/TBSignatureProfiler/issues}
}

}
\author{
\strong{Maintainer}: David Jenkins \email{dfj@bu.edu}

}
\keyword{internal}
