\name{tbsignatureprofilerapp}
\alias{tbsignatureprofilerapp}
\title{
Run the TBSignatureProfiler app
}
\description{
This function opens the TBSignatureProfiler R Shiny web application
}
\usage{
tbsignatureprofilerapp()
}


\value{
The R Shiny app will open
}

\author{
Christian Love
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
#Open the application
if(interactive()){
tbsignatureprofilerapp()
}
}

