% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TBsignaturesSplit}
\alias{TBsignaturesSplit}
\title{Up/Down-regulated genes information for selected TB signatures.}
\format{
list
}
\source{
See \code{?TBsignatures} for reference information.
}
\usage{
TBsignaturesSplit
}
\description{
Up/Down-regulated genes information for selected TB signatures.
}
\examples{
data("TBsignaturesSplit")
}
\keyword{datasets}
