% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSPapp.R
\name{TBSPapp}
\alias{TBSPapp}
\title{Run the TBSignatureProfiler Shiny application.}
\usage{
TBSPapp()
}
\value{
The Shiny application will open.
}
\description{
Use this function to run the TBSignatureProfiler application.
}
\examples{
# Upload data through the app
if (interactive()){
  TBSPapp()
  }

}
