% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{Bootstrap_LOOCV_LR_AUC}
\alias{Bootstrap_LOOCV_LR_AUC}
\title{Bootstrap LOOCV with logistic regression}
\usage{
Bootstrap_LOOCV_LR_AUC(df, target.vec, nboot)
}
\arguments{
\item{df}{a data.frame of gene expression count data.
Required.}

\item{target.vec}{binary vector indicating y}

\item{nboot}{number of bootstrap}
}
\value{
AUC from bootstrap LOOCV
}
\description{
Bootstrap LOOCV with logistic regression
}
\examples{
inputTest <- matrix(rnorm(1000), 100, 20,
                    dimnames = list(paste0("gene", 1:100),
                                    paste0("sample", 1:20)))
inputTest <- as.data.frame(inputTest)
targetVec <- sample(c(0,1), replace=TRUE, size=20)
nboot <- 3
res <- Bootstrap_LOOCV_LR_AUC(inputTest, targetVec, nboot)
}
