% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{deseq2_norm_rle}
\alias{deseq2_norm_rle}
\title{Normalize Gene Expression Count Data.}
\usage{
deseq2_norm_rle(inputData)
}
\arguments{
\item{inputData}{a \code{data.frame} or \code{matrix} of gene expression
count data. Required.}
}
\value{
A \code{data.frame} or \code{matrix} of normalized count data.
}
\description{
Normalize Gene Expression Count Data.
}
\examples{
## Example using the counts assay from a SummarizedExperiment
data_in <- SummarizedExperiment::assay(TB_indian, "counts")
res <- deseq2_norm_rle(data_in)
}
