% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalModel.R
\name{evaluateOriginalModel}
\alias{evaluateOriginalModel}
\title{A function that implements TB gene signatures' original model.}
\usage{
evaluateOriginalModel(
  input,
  useAssay = 1,
  geneSignaturesName = NULL,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE)
)
}
\arguments{
\item{input}{A SummarizedExperiment object with gene symbols as the assay row names.}

\item{useAssay}{A character string or an integer specifying the assay in the \code{input}.
Used for the test SummarizedExperiment object. Default is 1, indicating the 
first assay in the \code{input}.}

\item{geneSignaturesName}{A character string/vector specifying the signature of interest.
Default is NULL. If geneSignaturesName is NULL, run all available gene signatures' original model in the package.}

\item{BPPARAM}{An instance inherited from \code{bplappy}. 
See \code{\link[BiocParallel]{bplapply}} for details.}
}
\value{
A SummarizedExperiment object with predicted scores for each sample obtained from the signature's original model.
}
\description{
TB signatures with available original models are: Anderson_42,
Anderson_OD_51, Kaforou_27, Kaforou_OD_44, Kaforou_OD_53, Sweeney_OD_3,
Maertzdorf_4, Verhagen_10, Jacobsen_3, Sambarey_HIV_10, Leong_24, Berry_OD_86,
Berry_393, Bloom_OD_144, Suliman_RISK_4, Zak_RISK_16, and Leong_RISK_29.
The predicted score for each signature has been stored in the column data section
of the input SummarizedExperiment study.
}
\examples{
library(TBSignatureProfiler)
data("TB_hiv",package = "TBSignatureProfiler")
re <- evaluateOriginalModel(input = TB_hiv, useAssay = "counts", 
geneSignaturesName = c("Anderson_42","Maertzdorf_15"))
re$Anderson_42_OriginalModel                                                
}
