% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{runTBsigProfiler}
\alias{runTBsigProfiler}
\title{Run TB Signature Profiling}
\usage{
runTBsigProfiler(input, useAssay = NULL, signatures = NULL,
  algorithm = c("GSVA", "ssGSEA", "ASSIGN"),
  combineSigAndAlgorithm = FALSE, assignDir, outputFormat = NULL,
  parallel.sz = 0)
}
\arguments{
\item{input}{Input data as a SummarizedExperiment, data.frame, or matrix.
Required}

\item{useAssay}{When using SummarizedExperiment style input, the assay to
use for signature profiling. Required for SummarizedExperiment}

\item{signatures}{A list of signatures to run. If signatures is NULL, the
default set of signatures from the included TBsignatures data set is run.
For details run ?TBsignatures.}

\item{algorithm}{A list of algorithms to run. The default is GSVA, ssGSEA,
and ASSIGN.}

\item{combineSigAndAlgorithm}{Logical. If TRUE, the row names will be in
the form <signature>_<algorithm>. If FALSE, there will be a column named
'algorithm' that lists which algorithm is used. If NULL, and one algorithm
was used, the algorithm will not be listed. The default is false.}

\item{assignDir}{If using ASSIGN, a directory to save intermediate ASSIGN
results. The default is NULL and intermediate results will not be saved.}

\item{outputFormat}{Output data format. Can be SummarizedExperiment, matrix,
or data.frame. The default is the same as input.}

\item{parallel.sz}{Number of processors to use when doing the calculations in
parallel for GSVA and ssGSEA. If 0, all cores are used. The default is 0.}
}
\value{
A data.frame of signature profiling results.
}
\description{
Run profiling using a set of signatures and profiling algorithms
}
