% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{signatureGeneHeatmap}
\alias{signatureGeneHeatmap}
\title{Plot a Heatmap of a single signature scores and gene expression}
\usage{
signatureGeneHeatmap(inputData, useAssay, sigGenes, annotationData,
  name = "Signature", signatureColNames, annotationColNames,
  scale = TRUE, showColumnNames = TRUE, showRowNames = TRUE,
  colorSets = c("Set1", "Set2", "Set3", "Pastel1", "Pastel2", "Accent",
  "Dark2", "Paired"))
}
\arguments{
\item{inputData}{Either a SummarizedExperiment object that contains the
signature data and annotation as colData columns, or a data.frame or matrix
of signature data. Required.}

\item{useAssay}{If inputData is a SummarizedExperiment, the assay to use for
the gene expression data. Required if inputData is SummarizedExperiment.}

\item{sigGenes}{The genes in the signature to use in the heatmap. For inbuilt
signatures, you can use TBsignatures, (ex: TBsignatures[["ACS_COR"]]).
Required.}

\item{annotationData}{If inputData is a data.frame or matrix of signature
data, a data.frame or matrix of annotation data.}

\item{name}{The name of the heatmap. The default is "Signature".}

\item{signatureColNames}{If inputData is a SummarizedExperiment, the column
names in colData that contain the signature data.}

\item{annotationColNames}{If inputData is a SummarizedExperiment, the column
names in colData that contain the annotation data.}

\item{scale}{Scale the gene expression data. The default is TRUE.}

\item{showColumnNames}{Show columns names (sample names) in the heatmap. The
default is TRUE.}

\item{showRowNames}{Show row names(signature names) in the heatmap. The
default is TRUE.}

\item{colList}{Custom color information for annotation data as a named list
to pass to ComplexHeatmap. By default, ColorBrewer color sets will be used.
See colorSets for additional options.}
}
\description{
Plot a Heatmap of a single signature scores and gene expression
}
