% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{SignatureQuantitative}
\alias{SignatureQuantitative}
\title{Use logistic regression and Bootstrap LOOCV for signature evaluation.}
\usage{
SignatureQuantitative(df.input, target.vec.num, signature.list,
  signature.name.vec, num.boot = 100)
}
\arguments{
\item{df.input}{a data.frame of gene expression count data. 
Required.}

\item{target.vec.num}{numeric binary vector indicating y}

\item{signature.list}{a list of signatures}

\item{signature.name.vec}{vector of signature names}

\item{num.boot}{number of bootstrap}
}
\value{
the AUC, sensitivity and specificity
}
\description{
Use logistic regression and Bootstrap LOOCV for signature evaluation.
}
\examples{
inputTest <- matrix(rnorm(1000), 100, 10, 
                    dimnames = list(paste0("gene", 1:100),
                                    paste0("sample", 1:10)))
inputTest <- as.data.frame(inputTest)   
targetVec <- sample(c(0,1), replace=TRUE, size=10) 
signature.list <- list(sig1 = c("gene1", "gene2", "gene3")，
                       sig2 = c("gene4", "gene5", "gene6"))  
signature.name.vec <- c("sig1", "sig2")                       
num.boot <- 50                                   
res <- SignatureQuantitative(inputTest, 
                             targetVec, 
                             signature.list, 
                             signature.name.vec,
                             num.boot)
}
