% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TB_indian}
\alias{TB_indian}
\title{TB toy dataset}
\format{SummarizedExperiment}
\usage{
TB_indian
}
\description{
The gene expression and metadata in a SummarizedExperiment object.
}
\examples{
data("TB_indian")
}
\keyword{datasets}
