% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TBsignatures}
\alias{TBsignatures}
\title{TB Signatures}
\format{list}
\source{
TODO
}
\usage{
TBsignatures
}
\description{
A set of 15 TB signatures from publications. This set of signatures uses
gene symbols. Additional sets for entrez IDs and ensembl IDs are forthcoming.
}
\examples{
data("TBsignatures")
}
\keyword{datasets}
